requirements.txt
aiogram>=3.0
aiohttp
aiosqlite
python-dotenv
